requirements.txt
torch>=2.6.0
torchvision>=0.21.0
opencv-python>=4.10.0.84
opencv-contrib-python>=4.10.0.84
timm>=1.0.7
numpy>=1.26.4,<3
pyyaml>=6.0.1
tqdm>=4.66.4
tensorboard>=2.17.0
scikit-image>=0.24.0
Pillow>=10.4.0
h5py>=3.11.0
